deprecate storybook-preset and vestiges of per-app storybooks
