+++ apps/storybook/.storybook/main.ts
@@ -1,7 +1,5 @@
 import type { StorybookConfig } from '@storybook/react-vite'
 
-import { storykitAddons, framework } from '@priceline/storybook-preset'
-
 const packagesWithStories = [
   'carousel',
   'core',
@@ -14,13 +12,23 @@ const packagesWithStories = [
 ]
 const storiesGlob = 'src/**/*.stories.@(js|jsx|ts|tsx)'
 
+const stories = packagesWithStories.map(
+  (packageFolder) => `../../../packages/${packageFolder}/${storiesGlob}`
+)
+
 const config: StorybookConfig = {
-  framework,
-  stories: packagesWithStories.map((packageFolder) => `../../../packages/${packageFolder}/${storiesGlob}`),
-  addons: [...storykitAddons, '@storybook/addon-interactions'],
+  addons: [
+    '@storybook/addon-essentials',
+    '@storybook/addon-a11y',
+    '@storybook/addon-interactions',
+    'storybook-addon-designs',
+    '@storybook/addon-coverage',
+  ],
   docs: {
     autodocs: true,
   },
+  framework: '@storybook/react-vite',
+  stories,
   typescript: {
     check: false,
   },
