+++ apps/storybook/package.json
@@ -13,32 +13,37 @@
     "storybook:baseline": "run-s storybook:build storybook:baseline:chromatic",
     "storybook:build": "storybook build --output-dir=../docs/out/storybook",
     "storybook:baseline:chromatic": "chromatic --storybook-build-dir=../docs/out/storybook --auto-accept-changes --exit-once-uploaded",
-    "chromatic": "chromatic --storybook-build-dir=../docs/out/storybook --exit-once-uploaded --only-changed --trace-changed --untraced=**/package.json"
+    "chromatic": "chromatic --storybook-build-dir=../docs/out/storybook --exit-once-uploaded --only-changed --trace-changed --untraced=**/package.json",
+    "storybook:coverage": "test-storybook --coverage"
   },
   "author": "Priceline Design System Working Group",
   "license": "MIT",
   "devDependencies": {
     "@priceline/eslint-config": "workspace:*",
-    "@priceline/storybook-preset": "workspace:*",
-    "@rushstack/eslint-patch": "^1.3.1",
-    "@storybook/addon-actions": "^7.0.20",
-    "@storybook/addon-a11y": "^7.0.20",
-    "@storybook/addon-essentials": "^7.0.20",
-    "@storybook/addon-interactions": "^7.0.26",
-    "@storybook/react": "^7.0.20",
-    "@storybook/react-vite": "^7.0.20",
+    "@rushstack/eslint-patch": "^1.3.2",
+    "@storybook/addon-actions": "^7.0.27",
+    "@storybook/addon-a11y": "^7.0.27",
+    "@storybook/addon-coverage": "^0.0.8",
+    "@storybook/addon-essentials": "^7.0.27",
+    "@storybook/addon-interactions": "^7.0.27",
+    "@storybook/addon-viewport": "^7.0.27",
+    "@storybook/jest": "^0.1.0",
+    "@storybook/react": "^7.0.27",
+    "@storybook/react-vite": "^7.0.27",
+    "@storybook/testing-library": "^0.2.0",
     "babel-plugin-react-docgen": "^4.2.1",
-    "chromatic": "^6.19.5",
-    "eslint": "^8.42.0",
+    "chromatic": "^6.19.9",
+    "eslint": "^8.45.0",
     "npm-run-all": "^4.1.5",
     "pcln-design-system": "workspace:*",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-is": "^17.0.2",
     "require-from-string": "2.0.2",
-    "storybook": "^7.0.20",
+    "storybook": "^7.0.27",
+    "storybook-addon-designs": "beta",
     "styled-components": "^5.3.11",
-    "vite": "^4.3.9",
-    "typescript": "^5.1.3"
+    "vite": "^4.4.4",
+    "typescript": "^5.1.6"
   }
 }
