+++ common/changes/pcln-popover/cleanup-and-viteup_2023-07-15-03-39.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "comment": "Delete storybook-preset and vestiges of storybook configuration in individual projects, apps/storybook now starts in 1 second",
+      "type": "patch",
+      "packageName": "pcln-popover"
+    }
+  ],
+  "packageName": "pcln-popover",
+  "email": "steven.dalonzo@priceline.com"
+}
