+++ packages/carousel/package.json
@@ -54,8 +54,7 @@
     "@priceline/babel-preset": "workspace:*",
     "@priceline/eslint-config": "workspace:*",
     "@rushstack/eslint-patch": "^1.3.1",
-    "@storybook/addon-actions": "^7.0.19",
-    "@storybook/react": "^7.0.19",
+    "@storybook/addon-actions": "^7.0.27",
     "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^12.1.2",
     "@testing-library/react-hooks": "^7.0.2",
