+++ packages/core/package.json
@@ -89,19 +89,18 @@
     "@priceline/babel-preset": "workspace:*",
     "@priceline/eslint-config": "workspace:*",
     "@priceline/heft-component-rig": "workspace:*",
-    "@priceline/storybook-preset": "workspace:*",
     "@typescript-eslint/parser": "5.59.9",
     "@typescript-eslint/eslint-plugin": "5.59.9",
     "@rushstack/eslint-patch": "^1.3.1",
     "@rushstack/eslint-config": "^3.3.1",
     "@rushstack/heft": "~0.50.7",
-    "@storybook/addon-actions": "^7.0.20",
-    "@storybook/addon-docs": "^7.0.20",
-    "@storybook/addon-links": "^7.0.20",
+    "@storybook/addon-actions": "^7.0.27",
+    "@storybook/addon-docs": "^7.0.27",
+    "@storybook/addon-interactions": "^7.0.27",
+    "@storybook/addon-links": "^7.0.27",
     "@storybook/jest": "^0.1.0",
-    "@storybook/react": "^7.0.20",
-    "@storybook/addon-interactions": "^7.0.20",
-    "@storybook/testing-library": "^0.1.0",
+    "@storybook/react": "^7.0.27",
+    "@storybook/testing-library": "^0.2.0",
     "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^12.1.2",
     "@testing-library/react-hooks": "^7.0.2",
