+++ packages/modal/package.json
@@ -33,8 +33,8 @@
     "@babel/core": "^7.16.7",
     "@priceline/babel-preset": "workspace:*",
     "@priceline/eslint-config": "workspace:*",
-    "@priceline/storybook-preset": "workspace:*",
     "@rushstack/eslint-patch": "^1.1.0",
+    "@storybook/addon-actions": "^7.0.27",
     "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^12.1.2",
     "cross-env": "^7.0.3",
