+++ packages/popover/package.json
@@ -39,7 +39,7 @@
     "@priceline/eslint-config": "workspace:*",
     "@reach/component-component": "^0.16.0",
     "@rushstack/eslint-patch": "^1.1.0",
-    "@storybook/addon-actions": "^7.0.20",
+    "@storybook/addon-actions": "^7.0.27",
     "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^12.1.2",
     "cross-env": "^7.0.3",
