--- packages/slider/.storybook/main.js
@@ -1,9 +1,0 @@
-const { heftStoryGlobs, storykitAddons, features, framework, core } = require('@priceline/storybook-preset')
-
-module.exports = {
-  stories: heftStoryGlobs,
-  addons: storykitAddons,
-  features,
-  framework,
-  core,
-}
