--- packages/slider/.storybook/preview-body.html
@@ -1,3 +1,0 @@
-<!-- Google Fonts (Montserrat) -->
-<link href="https://fonts.googleapis.com/css?family=Montserrat:500,700" rel="stylesheet" />
-<!-- End Google Fonts (Montserrat) -->
