+++ packages/slider/package.json
@@ -21,7 +21,6 @@
     "build-only": "heft build --clean",
     "debug": "heft --debug test --clean",
     "start": "heft start",
-    "storybook": "heft start --storybook",
     "test:snapshots": "heft test --update-snapshots",
     "test:watch": "heft test --watch"
   },
@@ -40,13 +39,10 @@
     "@babel/core": "^7.22.1",
     "@priceline/eslint-config": "workspace:*",
     "@priceline/heft-component-rig": "workspace:*",
-    "@priceline/storybook-preset": "workspace:*",
     "@reach/component-component": "^0.17.0",
     "@rushstack/eslint-config": "^3.3.1",
     "@rushstack/eslint-patch": "^1.3.1",
     "@rushstack/heft": "~0.50.7",
-    "@storybook/addon-actions": "^7.0.18",
-    "@storybook/react": "^7.0.18",
     "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^12.1.2",
     "@types/jest": "^29.4.0",
