+++ rigs/component/package.json
@@ -11,7 +11,6 @@
     "@microsoft/api-extractor": "~7.35.2",
     "@rushstack/eslint-config": "^3.3.1",
     "@rushstack/heft-jest-plugin": "~0.5.13",
-    "@rushstack/heft-storybook-plugin": "~0.2.12",
     "@rushstack/heft-web-rig": "~0.15.2",
     "@rushstack/heft-webpack5-plugin": "~0.6.12",
     "@types/heft-jest": "^1.0.3",
