--- tools/storybook-preset/package.json
@@ -1,93 +1,0 @@
-{
-  "name": "@priceline/storybook-preset",
-  "version": "0.0.1",
-  "description": "Storybook Preset with commonly used addons and tools",
-  "scripts": {
-    "_phase:build": "heft build --clean",
-    "_phase:test": "heft test --no-build",
-    "build": "heft test --clean",
-    "build-only": "heft build --clean",
-    "debug": "heft --debug test --clean",
-    "postpublish": "node ../../common/scripts/install-run-rush.js slack-publish-notification --package @pcln/storybook-preset",
-    "test:snapshots": "heft test --no-build --update-snapshots",
-    "test:watch": "heft test --watch"
-  },
-  "main": "lib-commonjs/index.js",
-  "module": "lib-esm/index.js",
-  "types": "types/index.d.ts",
-  "license": "UNLICENSED",
-  "author": "Steve Dalonzo <steven.dalonzo@priceline.com>",
-  "keywords": [
-    "storybook",
-    "documentation"
-  ],
-  "repository": {
-    "directory": "storybook/storybook-preset",
-    "type": "git",
-    "url": "github:pcln/pcln-web"
-  },
-  "files": [
-    "lib-esm",
-    "lib-commonjs",
-    "types"
-  ],
-  "publishConfig": {
-    "registry": "https://npm.prod.pcln.com/repository/npm-pcln/"
-  },
-  "dependencies": {
-    "@babel/core": "^7.22.5",
-    "@babel/plugin-transform-react-jsx": "^7.22.5",
-    "@babel/runtime": "^7.22.5",
-    "@storybook/addon-a11y": "^7.0.20",
-    "@storybook/addon-actions": "^7.0.20",
-    "@storybook/addon-essentials": "^7.0.20",
-    "@storybook/addon-links": "^7.0.20",
-    "@storybook/addon-viewport": "^7.0.20",
-    "@storybook/addons": "^7.0.20",
-    "@storybook/api": "^7.0.20"
-  },
-  "devDependencies": {
-    "@priceline/eslint-config": "workspace:*",
-    "@priceline/heft-component-rig": "workspace:*",
-    "@rushstack/eslint-config": "~3.3.1",
-    "@rushstack/heft": "~0.50.7",
-    "@storybook/addon-docs": "^7.0.20",
-    "@storybook/cli": "^7.0.20",
-    "@storybook/components": "^7.0.20",
-    "@storybook/core-events": "^7.0.20",
-    "@storybook/react": "^7.0.20",
-    "@storybook/react-vite": "^7.0.20",
-    "@storybook/theming": "^7.0.20",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.1",
-    "@types/node": "^18.16.19",
-    "@types/react": "^17.0.62",
-    "@types/react-dom": "^17.0.20",
-    "@types/webpack-env": "^1.18.1",
-    "css-loader": "~6.8.1",
-    "eslint": "^8.42.0",
-    "pcln-design-system": "workspace:*",
-    "pcln-icons": "workspace:*",
-    "pcln-menu": "workspace:*",
-    "pcln-popover": "workspace:*",
-    "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-is": "^17.0.2",
-    "require-from-string": "2.0.2",
-    "source-map-loader": "^4.0.1",
-    "storybook-addon-performance": "^0.17.1",
-    "storybook-addon-designs": "beta",
-    "style-loader": "^3.3.3",
-    "styled-components": "^5.3.11",
-    "typescript": "^5.1.3",
-    "vite": "^4.3.9",
-    "webpack": "^5.86.0"
-  },
-  "peerDependencies": {
-    "pcln-design-system": "^5",
-    "pcln-icons": "^5.6.0",
-    "react": ">=16.12.0",
-    "react-dom": ">=16.12.0"
-  }
-}
