--- tools/storybook-preset/src/__test__/index.ts
@@ -1,1 +1,0 @@
-export * from './testing-library'
