--- tools/storybook-preset/src/__test__/jest.setup.ts
@@ -1,2 +1,0 @@
-import 'jest-styled-components'
-import '@testing-library/jest-dom/extend-expect'
