--- tools/storybook-preset/src/__test__/testing-library.tsx
@@ -1,13 +1,0 @@
-import React, { FC, ReactElement } from 'react'
-import { render, RenderOptions, RenderResult } from '@testing-library/react'
-import { ThemeProvider } from 'pcln-design-system'
-
-const AllTheProviders: FC = ({ children }) => {
-  return <ThemeProvider>{children}</ThemeProvider>
-}
-
-const customRender = (ui: ReactElement, options?: Omit<RenderOptions, 'wrapper'>): RenderResult =>
-  render(ui, { wrapper: AllTheProviders, ...options })
-
-export * from '@testing-library/react'
-export { customRender as render }
