--- tools/storybook-preset/src/parameters/index.ts
@@ -1,15 +1,0 @@
-import { a11y } from './a11y'
-import { viewports } from '../viewports'
-
-/** @public */
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-  a11y,
-  viewport: { viewports },
-}
