--- tools/storybook-preset/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "./node_modules/@priceline/heft-component-rig/profiles/react/tsconfig-base.json"
-}
